fix : 버그 수정
