+++ src/utils/api/http.js
@@ -1,11 +1,27 @@
 import axios from "axios";
 
-axios.defaults.headers.common["access-token"] = sessionStorage.getItem("access-token");
-
-export default axios.create({
+const instance = axios.create({
   baseURL: "http://localhost",
   headers: {
     "Content-Type": "application/json;charset=utf-8",
-    Authorization: sessionStorage.getItem("access-token"),
   },
 });
+
+instance.interceptors.request.use(
+  (config) => {
+    const accessToken = sessionStorage.getItem("access-token");
+
+    if (accessToken) {
+      config.headers["Authorization"] = accessToken;
+      config.headers["access-token"] = accessToken;
+    }
+    return config;
+  },
+  (error) => {
+    return Promise.reject(error);
+  }
+);
+
+export default instance;
+
+// 이거 axios export 대문에 에러나는거니까 오후에 와서해 인혁아
