+++ src/views/home/TheIndex.vue
@@ -82,17 +82,48 @@
 
 <script>
 import { BIcon } from "bootstrap-vue";
+import jwtDecode from 'jwt-decode';
+import http from '@/utils/api/http';
+import { mapActions, mapState } from 'vuex';
 
 export default {
   name: "TheIndex",
   components: { BIcon },
   data() {
     return {
-      message: "",
     };
   },
-  created() {},
-  methods: {},
+  created() {
+    console.log(this.isLogin)
+    console.log(sessionStorage.getItem("access-token"))
+    if (this.isLogin) {
+        this.getMemberInfo();
+    }
+  },
+  computed: {
+    ...mapState("memberStore",["isLogin"])
+  },
+  methods: {
+    ...mapActions("memberStore",["setLoginMember"]),
+    getMemberInfo() {
+      const accessToken = sessionStorage.getItem("access-token");
+      const decodedAccessToken = jwtDecode(accessToken);
+      const memberId = decodedAccessToken.memberId;
+      
+       http
+        .get(`/member/info/${memberId}`)
+        .then((res) => {
+          console.log(res)
+          this.setLoginMember(res.data.loginMember);
+        })
+        .catch((error) => {
+          console.log(error);
+          this.$alertDanger(
+            "사용자 정보 불러오기 실패 !",
+            "추후 예외처리 추가 예정"
+          );
+        });
+    },},
 };
 </script>
