+++ src/views/member/MemberLogin.vue
@@ -58,7 +58,6 @@
 
 <script>
 import http from "@/utils/api/http";
-import jwtDecode from "jwt-decode";
 import { mapActions, mapState } from "vuex";
 
 export default {
@@ -78,8 +77,8 @@ export default {
   },
   methods: {
     ...mapActions("memberStore", ["setIsLogin", "setLoginMember"]),
-    async login() {
-      await http
+    login() {
+      http
         .post("/member/login", this.loginMember)
         .then((res) => {
           console.log(res);
@@ -103,29 +102,9 @@ export default {
           this.setIsLogin(false);
         });
 
-      if (this.isLogin) {
-        this.getMemberInfo();
-      }
+      
     },
-    async getMemberInfo() {
-      const accessToken = sessionStorage.getItem("access-token");
-      const decodedAccessToken = jwtDecode(accessToken);
-      const memberId = decodedAccessToken.memberId;
-
-      await http
-        .get(`/member/info/${memberId}`)
-        .then((res) => {
-          console.log(res)
-          this.setLoginMember(res.data.loginMember);
-        })
-        .catch((error) => {
-          console.log(error);
-          this.$alertDanger(
-            "사용자 정보 불러오기 실패 !",
-            "추후 예외처리 추가 예정"
-          );
-        });
-    },
+    
   },
 };
 </script>
